Added handling for a pre-existing .bashrc file

+++ provision.sh
@@ -25,7 +25,21 @@ fi
 # Vim
 ln -s "./dotfiles/.vimrc" "$HOME/.vimrc"
 # Bash
-ln -s "./dotfiles/.bashrc" "$HOME/.bashrc"
+
+if [ ! -f "$HOME/.bashrc" ]; then
+    # .bashrc files are sacred; Don't bloody touch them without explicit permission.
+    # If this script is being used, it's likely that the pre-existing .bashrc is just the distro default.
+    read -rp "A .bashrc file already exists; Would you like to overwrite it? [y/n]: " confirm 
+    if [[ $confirm == [yY] || $confirm == [yY][eE][sS] ]]; then
+	rm "$HOME/.bashrc"
+	ln -s "./dotfiles/.bashrc" "$HOME/.bashrc"
+    else
+	echo "Leaving the pre-existing .bashrc alone."
+    fi
+else
+    ln -s "./dotfiles/.bashrc" "$HOME/.bashrc"
+fi
+
 ln -s "./dotfiles/.bash_profile" "$HOME/.bash_profile"
 ln -s "./dotfiles/.ps1_prompt" "$HOME/.ps1_prompt"
 # Htop
